switch index
make demos the index
